[UPM][UPMErrorUI] Wire the button in the password manager error message
The "Sign in" button starts the Android sign in process using the AccountManager.
PasswordManagerSignInHelperBridge.java is added so the process can be
started form the cpp side.
The password_manager_sign_in_helper_bridge.h is an interface added for
the purpose of tests.
There is a real implementation calling the free function startSignInFlow
and a mock implementation so we can check the method was called in
tests.

Bug: 1352415
Change-Id: I8fb213d2de656d6e5137b8eb10f39a1cebd4d714
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3855448
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1045029}