Add backward migration policy.
Decide how to migrate back when lacros is disabled. This is still a prototype and we might decide to change the functionality later based on  how well it works.

Policy will be used in a future CL to trigger backward migration.

DD: go/lacros-ash-migration-plan

Bug: b:244572632
Change-Id: I763d10a7dafe8932cb57891beba853183acc8f15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3836817
Reviewed-by: Jana Grill <janagrill@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045037}