Separate AggregationServiceStorage::RequestId type definition
Moves the definition of that ID type to its own header to allow mojo
traits files to depend on that single type only. Currently, a larger
header is required, which can cause flaky build issues as other headers
are (transitively) #included that are not listed in the deps.

This cl does not rename existing usage to minimize churn.

Bug: 1361469
Change-Id: Id6ab6d4035448a2bc68370016a32ef912ac31ca7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3876647
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Auto-Submit: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045116}