Fix DCHECK when restoring from bfcache several times in a row
ContentToVisibleTimeReporter was not clearing state when a frame was
shown for bfcache restore without a tab switch. This triggered a
DCHECK the next time the frame was shown as the stale state still
existed.

Re-enables Side Search tests for Mac that triggered the check.

Bug: 1348296
Change-Id: I3a3cc685d5b5511460a8e2774d539b8df90ace95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3884322
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045124}