NetworkPortalNotificationController: PortalSuspected notification
This updates the system notification generated by the
NetworkPortalNotificationController to have a different message and
button title for the case of the portal state being 'portal-suspected'.

Browser tests have also been added to verify the title, message, and
button title for different portal states and when the captive portal ui
flag is on or off.

BUG=b:242730393
TEST=unit_tests --gtest_filter="NetworkPortalNotificationControllerTest*"
browser_tests --gtest_filter="NetworkPortalDetectorImplBrowserTest*"

Change-Id: I16833e7db87a8eec7abe8a9adcddb0fa2cc702d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3872589
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Michael Rygiel <michaelrygiel@google.com>
Cr-Commit-Position: refs/heads/main@{#1045143}