[Passwords] Expose parse status in CSVPassword mojo.
Exposing the parse status of a CSV row in the mojo interface. This information is required when a row is not parsed correctly.

LOW_COVERAGE_REASON=Coverage will increase once the symantic/syntax errors will be used in importer flow in a follow up cl.

Bug: 1325290
Change-Id: I2b414a75d196ec8c0806adf6af8503f31be18f0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3879619
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Elias Khsheibun <eliaskh@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045178}