[Test Automation] Change timing of drag cancel to decrease flakiness
After some research, I found that the flakes were due to hangs in the
test code where the system got caught in a delayed drag-drop runloop
after the final test step but before the sequence completed.

Therefore, the pre-emptive drag cancel we put in before to avoid a hang
between the end of the test body and the test teardown had to be moved
earlier; to the last step of the test.

This CL pulls out drag-drop specific code into its own test fixture,
simplifying it, and moves the cancel up to the final step of the test,
after the result of the test has been confirmed.

In a follow-up CL, a generalized drag-drop action will be created for
interactive testing that automatically does cleanup on any platform that
requires this type of cleanup.

Bug: 1354311
Change-Id: Ib4a2b818385a5d6ef458cadc04736eb040008b6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3866147
Reviewed-by: Connie Wan <connily@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045194}