saved_desks: Long press template to commit name changes.
This CL fixed the issue that when long press on a template card which
has been clicked and in edit mode, it doesn't have any response.

After the fix, we will commit the name changes for the template and show
the hover button. Also, when we long press on another template card, we
will also commit the name change. For scroll press, as Gary said, we
should commit name changes, but don't update hover button visibility.

Fixed: 1344942
Change-Id: I2b147eb3d04119a78e463564c21a7c84e8333ff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3867971
Commit-Queue: Hongyu Long <hongyulong@chromium.org>
Reviewed-by: Daniel Andersson <dandersson@chromium.org>
Reviewed-by: Yongshun Liu <yongshun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045218}