Handle tab closing mode for tabs within a group.
In tab closing mode, the parameter for the override width would reset the value to be nullopt which would be an issue for group tabs as this would change the value of the available_space_for_tabs not respecting the tab closing mode.

LOW_COVERAGE_REASON=For tab closing, we do not pass a non null value. Also, otherwise it just takes the argument value and is trivial. Validated that the behaviour for tab group toggling.

Bug: 1322791
Change-Id: I1144ad0041025d5beba7ca6f87719940fe2d0779
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3869188
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Shibalik Mohapatra <shibalik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045248}