[GridNG] Mitigate perf regression from NGGridLineResolver constructor
https://chromium-review.googlesource.com/c/chromium/src/+/3816648 added
a constructor to NGGridLineResolver. This constructor copies some
WTF::HashMaps (via NamedGridLinesMap and ComputedGridTrackList),
which can be expensive with very large grids (as in this test).

This mitgation avoids this expense by storing a scoped_refptr
to the ComputedStyle. This will be avoid the expensive copies
for non-subgrid's.

Pinpoint #'s (runs/second, higher is better):
Before regressing change:   8882.4731
After regressing change:    6816.3732
After this mitigation:      8,760.685

Bug: 1357258
Change-Id: I8d93a47c05f99fb2fdbab7fbf4bd1f1b8f068fff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3863591
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1045251}