Avoid Password Details Crash and add monitoring metrics
This patch works around a crash on the Password Details page. The crash
happens mostly on iOS 16 and is caused by a nil UIMenuController when
trying to copy a field to the pasteboard. The nil menu would lead to a
crash when building the confirmation Snackbar with a nil message.

This patch avoids showing the Snackbar if its would-be message is nil
and logs a histogram for monitoring the issue.

Bug: 1359331
Change-Id: Ib19aade7e42e65d3e950e7952230b8508231b18f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3884317
Auto-Submit: Ernesto Izquierdo Clua <eic@google.com>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Commit-Queue: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1045265}