personalization: hide aria-selected while image loading
Hide the checkmark indicating an item is selected until it loses
the placeholder attribute. Follow up to recent CL to keep the
placeholder CL set until image loading has resolved. It resulted
in the blue checkmark flashing in and out with the placeholder
animation, which looked weird.

BUG=b:242355439

Change-Id: Ia4c2a96cf3e397eebe4c85990f42ceb3e4e2c08c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3868187
Reviewed-by: Tao Wu <wutao@chromium.org>
Commit-Queue: Jeffrey Young <cowmoo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045285}