Ensure cert verifiers created by CertVerifierServiceFactory use the same implementation.
Previously the implementation to use was configured as part of CertVerifierCreationParams which is passed each time a cert verifier is created, meaning different implementations could be created unless every call site used the same code for configuring the params.

The new implementation moves those configuration choices to a new CertVerifierServiceParams that is passed in when the CertVerifierServiceFactoryImpl is created.

A new embedder API ContentBrowserClient::GetCertVerifierServiceParams is added to allow the embedder control the parameters passed to the factory.

Bug: 1352171
Change-Id: I9ea39ae5b3667db1f5f5ec5739d086d63ba20871
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3831846
Commit-Queue: Matt Mueller <mattm@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045308}