[Libcast-in-Chrome #6] Introduce OpenscreenRpcDispatcher
Patch:
This patch adds an OpenscreenRpcDispatcher implementation based
on the openscreen::cast::SenderSessionMessenger.

For more context, see the reference patch that includes this plus other
changes:
https://chromium-review.googlesource.com/c/chromium/src/+/3868508.

Coming up next:
1. Backing code to allow getting the audio encoder's bitrate will be
added to the audio RTP stream.

2. A new class called OpenscreenSessionHost will be added to wrap an
openscreen::cast::SenderSession.

3. Metrics need to be added to make sure we don't introduce major
regressions.

4. Once these pieces are in place, extensive manual testing and
improving of the automated test suite needs to occur.

Overall feature:
Libcast-In-Chrome is a project with the goal of having libcast be the
backing implementation of cast streaming in Chrome.

Bug: 1318499, b/159352836
Change-Id: I31697d1f3c0f23ab3bdea5d8da070d0229d9ab11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3867701
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045311}