[PM] Various fixes in BackgroundTabLoadingPolicy.
Use "GetPermissionResultForOriginWithoutContext" instead of
"GetPermissionStatusForCurrentDocument" to determine if a WebContents
has the notification permission. The former relies on the committed
URL of the primary main frame, which is empty when a WebContents is
passed to BackgroundTabLoadingPolicy to be loaded.

Do not add a page to the list of pages to load if it's already
loading. This happens for visible restored pages, which immediately
start loading on creation.

Bug: 1059341
Change-Id: I8088bf30a20a7cc3621215bb541d031f192b352f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3883819
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045312}