enable_libaom requires enable_av1_decoder
//media/video/software_video_encoder_test.cc requires an AV1 decoder for
the AV1 encoder tests, which are enabled when enable_libaom. Fail to
compile if a decoder is not enabled.

Also add a note about this relationship to the `enable_av1_decoder`
declaration. Other than the test mentioned above, the relationship isn't
strictly required, so no GN assert is added.

Bug: 1352700
Change-Id: Ib1c6052eec6db14ea7b4fbe63d48822492310a0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3887160
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: David Dorwin <ddorwin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045321}