[Chromecast] Change start threshold only when having valid session id.
This is to avoid underrun when playing silence buffer.

When underrun happens, rendering delay will be reset but the upper
pipeline is waiting for a valid rendering delay.

When the session ID is valid, there's a special logic to return a
unstable rendering delay to minimize the start latency.

Bug: internal b/233141758
Test: On device playback
Change-Id: Ia34b2550dd73ddd9f826347273a37d8b55ace023
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3887229
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Auto-Submit: Yuchen Liu <yucliu@chromium.org>
Commit-Queue: Kenneth MacKay <kmackay@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045342}