cros_next: fix wrong bg color of pill button
We changed the PillButton's background to a themed background which
could automatically update the color when the theme changes. However
if there is a user-defined background color, themed background will
overwrite the color in
ThemedRoundedRectBackground::OnViewThemeChanged with color generated
from predefined color id.

To make the background color changing convenient, we change back to
solid color background.

Since AshColorProvider is deprecated, we can also get color provider
after the button is added to a widget. We make following changes:
- Move setting horizontal alignment, focus ring, and background in
Init function. The Init function can only be called when the button
is added to a widget.
- Move updating background color, text color, and icon color in
saparate functions. When there is a client-defined color, we use it.
Otherwise, we use the color generated from the predefined color id
after the button is added to a widget.

Bug: b:243990071
Change-Id: If11e629d157458234c76847c05563568dd0931c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3877566
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045361}