Remove invalid DCHECK in PrefetchService.
This DCHECK could be triggered if a RFH was reused across a same site
navigation and both pages requested to prefetch the same URL. In this
case the existing prefetch in PrefetchService::all_prefetches_ must be
owned by the PrefetchService, otherwise it would have been deleted when
the PrefetchDocumentManger for the first page was deleted. Therefore we
can call PrefetchService::ResetPrefetch on the existing prefetch to
remove it from the PrefetchService, and then treat the new prefetch like
normal.

Bug: 1358692
Change-Id: If916f308db5c0d204a7f8c7582bba6d1074cc9a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3880874
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Max Curran <curranmax@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045373}