Document phone number types of autofill
The documentation of phone number fields is spread throughout the
"enum ServerFieldType" and pretty hard to follow because fields are
sorted by protobuf ID but not in a logical order. This CL summarizes
the relevant information in one place and hopefully makes it easier
to process.

Bug: 1007974
Change-Id: Id9693c29c7f62e204a2c3a7510f6b85d2a2e4e19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3870569
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045375}