Implement QS Catalog
Added quick_settings_catalog file for QS views. This cl adds the
QsButtonCatalogName for buttons in the qs view.

To avoid a super large cl size, this cl only applies the new buttons
in the TopShortcutsView.

`QuickSettingsButtonDelegate` is the base button delegate for each qs
button. It handles the UMA `OnButtonActivated` tracking and adds the
corresponding `ViewID` to the button.

Every button will need to be created by the
`QuickSettingsButtonDelegate` to be added to the qs view
(in this cl it's TopShortcutsView later will be the entire qs view).

Bug: 1357233
Change-Id: I66e25ed9021660bc560c146f557c18b3e5464ce1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3860026
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045393}