Respond to DelegatedSourceList cancellation
Updates the chrome picker to respond when a delegated source list is
dismissed without the user making a selection (either via an explicit
cancellation or due to an error). If there is only one pane visible,
then we dismiss the chrome picker. However, if the "Tabs" pane is an
option, the chrome picker will switch back to that pane as the "default"
tab. Note that in order to prevent a potentially confusing loop for the
user, we explicitly do not switch back to the previous tab, which could
*also* be driven a delegated source list, and thus cause a very similar
dialog to the one the user just dismissed to re-appear.

Fixed: 1351577
Change-Id: I3fb63625e0231709ff0e624b656c4daf2c4476d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3869220
Reviewed-by: Elad Alon <eladalon@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045411}