Preloading: Associate Preloading UKMs with Navigation UkmSourceId
Currently, we are associating Preloading UKMs with the
RenderFrameHost::GetPageUkmSourceId. With this logic we are
missing PrerenderPageLoad for the same SourceId. This is because
PrerenderPageLoad UKM uses activation navigation id whereas we used
the prerender navigation page Ukm source id.

To fix this, so that we can get all the data for PrerenderPageLoad
along with Preloading UKMs, we associate the UKMs with the primary
page navigated source id by using DidFinishNavigation instead of
PrimaryPageChanged to use NavigationHandle::GetNextPageUkmSourceId().

Bug: 1359149
Change-Id: I3c5ba0d4644ee191eb6fc896a63fb6aae27663b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3885223
Reviewed-by: Simon Pelchat <spelchat@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045468}