Get download verdict in CSBRR from download item user data.
Previously, the download verdict is obtained by remapping the danger
type back to the download verdict. However, this is not accurate if
the user has already bypassed the warning. The danger type at that
point is DOWNLOAD_DANGER_TYPE_USER_VALIDATED, which will be mapped to
the safe verdict.

In this CL, rename DownloadPingToken to DownloadProtectionData, and
extend it to store the original verdict. When constructing CSBRR,
obtain it directly from DownloadProtectionData.

Bug: 1358262
Change-Id: I2c2d50059517700be2bb59f58eeab2fce4af0184
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3884803
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1045498}