[TTFCC] Suppress the keyboard while the TTF surface is shown
This CL adds a new way to suppress the keyboard on the browser side. It
is achieved by setting the always_hide_ime bit to true on every TextInputStateChanged event while Touch To Fill surface is shown. The
use of existing suppression approach in the renderer was not an option, since the decision on whether to show TTF or not is made on the browser side, see [1].

RenderWidgetHostImpl now manages a list of SuppressShowingImeCallbacks,
and suppresses the keyboard if any of them returns true. Callbacks are
attached by ContentAutofillDrivers during their lifetime and return
false by default.

Finally, the keyboard is controlled by the bottom sheet UI by calling ContentAutofillDriver::SetShouldSuppressKeyboard with a boolean every
time the TTF surface visibility changes.

[1]https://docs.google.com/document/d/1f8kIrYg19qsgOBTrsJzwFSFbBgfnbMG3Li5NwJdvl2o/edit?resourcekey=0-kRnKWPjAzLhqVGYiM_7W4w#heading=h.tgzqveoy1nac

Bug: 1247698
Change-Id: Id3f4a5289aa4b3e3c013fb8ff8c38ae51df78eed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3757786
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Elizabeth Popova <lizapopova@google.com>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1045503}